Change f'' to str()
Keith pointed out a numpy bug that isn't affecting this particular notebook but
could hit if rounded_val is ever numpy.float32. Details in

https://github.com/numpy/numpy/issues/10645

but my take on reading that issue is that str(rounded_val) is preferable to
f'{rounded_val}'. I re-ran the make_table() cells of the notebook to ensure the
updated function runs correctly so the cell execution count is not continuous
in this commit.