Add xpersist to Chl and nutrient plots
totChl, NO3, PO4, and SiO3 are all written to xpersist_cache/3d_fields
(although totChl is actually surface values...). Note the first three are
averaged over time dimension, while totChl is 25 years worth of output so it
can be converted to climatology if needed. (I'm open to better names for
directories and files, this was just a first pass to get things working.)

Note that I also pushed the intake catalog read into cu.read_CESM_var() --
perhaps that makes it less general than we want future versions to be, but for
now it's a good way to avoid parsing the catalog un-necessarily.