Change grid_mask in scrip grid file
Taking the mask to be all missing values from the January dataset leads to
masking out much of the arctic that is actually ocean (and has observations
from summer months). Instead the mask should be grid points that are missing in
EVERY monthly dataset.

Also, I added plots of the ann, DJF, and JJA data after being mapped onto the
POP grid to verify more data in the arctic.