Add burial to sed and NO3 riv flux
Keith pointed out that I was missing NO3_RIV_FLUX and ponToSed in my imbalance
term. NO3_RIV_FLUX should be added to the other two river fluxes, and ponToSed
is a whole new row in the table.

Also reporting the imbalance to the nearest tenth, rather than as a whole
number.