Fixed time dimension in resample_ann()
In Make Timeseries.ipynb, I had replaced esmlab.resample() with a new function
because esmlab.resample() doesn't work with the current stack. However, the new
function was setting the time dimension equal to the year being averaged,
rather than the midpoint of the points... e.g. 1990 rather than 1990-07-01.
This was resulting in the Global Flux Table using the wrong slice of years.

This commit includes:

1. updating the time dimension in the dataset returned by resample_ann()
2. updating ann_avg_utils:_get_time_and_ensemble_mean() to print a warning if
   the number of years being averaged doesn't match what is expected (which
   will catch averaging 1991-2014 instead of 1990-2014, but won't catch
   averaging 1989-1999 instead of 1990-2000)
3. regenerating the Global Flux Table using datasets with the correct time
   dimension