Check which years are being averaged over
The previous commit ensured that the length of ds["time"] in
_get_time_and_ensemble_mean() matched the number of years we wanted to average
over; this commit modifies the check to ensure we are also averaging over the
correct years by replacing experiment_len with experiment_span, a dictionary of
tuples (start_year, end_year). For now, I kept the previous error message for
cases where len(ds["time"]) != end_year - start_year + 1, but also added a
check that ds["time"].values[0].year == start_year and
ds["time"].values[-1].year == end_year when the length of the array is correct.

This caught an issue with the opal flux for SSP5-8.5, where we were averaging
from 2091 - 2100 instead of 2090 - 2099. I re-created the xpersist cache file
with the proper time array, and now we are averaging over the correct years
(which resulted in a slight change in the value of the table, 69.8 Tmol / year
to 69.9)